Add missing docs
