+++ buoyant-examples/battery/src/view/charge.rs
@@ -88,7 +88,7 @@ fn port_power_row(port_name: &str, power: i32) -> impl View<ColorFormat, ()> + u
             .padding(Edges::All, 5)
             .flex_infinite_width(HorizontalAlignment::Trailing)
             .with_infinite_max_height()
-            .background(Alignment::default(), || {
+            .background(Alignment::default(), {
                 RoundedRectangle::new(5).foreground_color(match power {
                     p if p > 0 => color::BLUE,
                     p if p < 0 => color::GREEN,
@@ -98,7 +98,7 @@ fn port_power_row(port_name: &str, power: i32) -> impl View<ColorFormat, ()> + u
     ))
     .with_spacing(SPACING * 2)
     .padding(Edges::All, 3)
-    .background(Alignment::default(), || {
+    .background(Alignment::default(), {
         RoundedRectangle::new(8).foreground_color(color::SECONDARY_BACKGROUND)
     })
 }
