+++ src/lib.rs
@@ -3,15 +3,15 @@
 #[cfg(any(test, feature = "std"))]
 extern crate std;
 
+pub mod animation;
 pub mod environment;
 pub mod event;
 pub mod font;
+pub mod image;
 pub mod layout;
 pub mod primitives;
 pub mod render;
 pub mod render_target;
-// #[warn(missing_docs)]
-pub mod animation;
-pub mod image;
 pub mod surface;
+#[warn(missing_docs)]
 pub mod view;
