+++ src/view/shape/rectangle.rs
@@ -7,7 +7,11 @@ use crate::{
 
 use super::RoundedRectangle;
 
-/// A rectangle
+/// A rectangle which takes space greedily on both axes.
+///
+/// By default, this renders a filled shape with the inherited foreground color.
+/// To render with a stroke instead, use [`ShapeExt::stroked`][`super::ShapeExt::stroked`]
+/// or [`ShapeExt::stroked_offset`][`super::ShapeExt::stroked_offset`].
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Default)]
 pub struct Rectangle;
