+++ src/view/shape/rounded_rectangle.rs
@@ -5,12 +5,18 @@ use crate::{
     view::{ViewLayout, ViewMarker},
 };
 
+/// A rounded rectangle which takes space greedily on both axes.
+///
+/// By default, this renders a filled shape with the inherited foreground color.
+/// To render with a stroke instead, use [`ShapeExt::stroked`][`super::ShapeExt::stroked`]
+/// or [`ShapeExt::stroked_offset`][`super::ShapeExt::stroked_offset`].
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Default)]
 pub struct RoundedRectangle {
     corner_radius: u16,
 }
 
 impl RoundedRectangle {
+    #[allow(missing_docs)]
     #[must_use]
     pub const fn new(corner_radius: u16) -> Self {
         Self { corner_radius }
